Rename squeezebox_player -> slimproto
