+++ homeassistant/components/slimproto/__init__.py
@@ -1,4 +1,4 @@
-"""Squeezebox Player integration."""
+"""SlimProto Player integration."""
 from __future__ import annotations
 
 from aioslimproto import SlimServer
