+++ homeassistant/components/slimproto/const.py
@@ -1,0 +1,8 @@
+"""Constants for SlimProto Player integration."""
+
+
+DOMAIN = "slimproto"
+
+DEFAULT_NAME = "SlimProto Player"
+
+PLAYER_EVENT = f"{DOMAIN}_event"
