+++ homeassistant/components/slimproto/strings.json
