+++ homeassistant/components/slimproto/translations/en.json
