--- homeassistant/components/squeezebox_player/const.py
@@ -1,8 +1,0 @@
-"""Constants for Squeezebox Player integration."""
-
-
-DOMAIN = "squeezebox_player"
-
-DEFAULT_NAME = "Squeezebox Player"
-
-PLAYER_EVENT = f"{DOMAIN}_event"
