+++ requirements_test_all.txt
@@ -206,7 +206,7 @@ aiosenz==1.0.0
 # homeassistant.components.shelly
 aioshelly==2.0.0
 
-# homeassistant.components.squeezebox_player
+# homeassistant.components.slimproto
 aioslimproto==1.0.0
 
 # homeassistant.components.steamist
