+++ tests/components/slimproto/conftest.py
@@ -1,4 +1,4 @@
-"""Fixtures for the Squeezebox Player integration tests."""
+"""Fixtures for the SlimProto Player integration tests."""
 from __future__ import annotations
 
 from collections.abc import Generator
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from homeassistant.components.squeezebox_player.const import DOMAIN
+from homeassistant.components.slimproto.const import DOMAIN
 
 from tests.common import MockConfigEntry
 
@@ -25,7 +25,7 @@ def mock_config_entry() -> MockConfigEntry:
 def mock_setup_entry() -> Generator[AsyncMock, None, None]:
     """Mock setting up a config entry."""
     with patch(
-        "homeassistant.components.squeezebox_player.async_setup_entry",
+        "homeassistant.components.slimproto.async_setup_entry",
         return_value=True,
     ) as mock_setup:
         yield mock_setup
