+++ tests/components/slimproto/test_config_flow.py
@@ -1,7 +1,7 @@
-"""Test the Squeezebox Player config flow."""
+"""Test the SlimProto Player config flow."""
 from unittest.mock import AsyncMock
 
-from homeassistant.components.squeezebox_player.const import DEFAULT_NAME, DOMAIN
+from homeassistant.components.slimproto.const import DEFAULT_NAME, DOMAIN
 from homeassistant.config_entries import SOURCE_USER
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import RESULT_TYPE_ABORT, RESULT_TYPE_CREATE_ENTRY
@@ -27,7 +27,7 @@ async def test_already_configured(
     mock_config_entry: MockConfigEntry,
     mock_setup_entry: AsyncMock,
 ) -> None:
-    """Test abort if Squeezebox Player is already configured."""
+    """Test abort if SlimProto Player is already configured."""
     mock_config_entry.add_to_hass(hass)
 
     result = await hass.config_entries.flow.async_init(
