--- tests/components/squeezebox_player/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Squeezebox Player integration."""
